@@ -1,8 +1,8 @@
 # Be sure to restart your server when you modify this file.
 
-WwwStudentpeaceprizeOrg::Application.config.session_store :cookie_store, :key => '_www.studentpeaceprize.org_session'
+WwwStudentpeaceprizeOrg::Application.config.session_store :cookie_store, key: '_www_studentpeaceprize_org_session'
 
 # Use the database for sessions instead of the cookie-based default,
 # which shouldn't be used to store highly confidential information
-# (create the session table with "rake db:sessions:create")
+# (create the session table with "rails generate session_migration")
 # WwwStudentpeaceprizeOrg::Application.config.session_store :active_record_store
