@@ -4,4 +4,4 @@
 # If you change this key, all old signed cookies will become invalid!
 # Make sure the secret is at least 30 characters and all random,
 # no regular words or you'll be exposed to dictionary attacks.
-WwwStudentpeaceprizeOrg::Application.config.secret_token = '706fd1c1ee233bf138d38c5b6f47cdf1764ad5a99d11bcbae8efa2f26af8a5e611e37f8caae37ccf74797565e3860a4cc017872e4f0add3b743766e371787fd6'
+WwwStudentpeaceprizeOrg::Application.config.secret_token = '493e00d55cc2aeff46c1a0660ae91ef96fecbef869e04537eca191da8c23f53ecf81b43947ad28da71794463635e91ddce70b7f7e724fb90358110eec5372218'
